Update README and Liceses
I've taken the text found in the index.md and added it to the README for the project.

I've also added a License section to the read that specifies the licenses in use for the project. The project itself will remain under MIT but the content is specified to be under CC-BY-4.0.

Signed-off-by: Froilan Irizarry Rivera <1918027+froi@users.noreply.github.com>